Add files via upload
The goal of this project is to develop an application that enables users to capture, analyze, andvisualize network packet statistics on a Linux system. The application will provide a command-line interface for users to specify capture filters, display preferences, and control the application's behavior. It will support real-time updating of packet statistics and offer both tabular and text-based graphical representations for visualizing network traffic.